perf test: Make each test/suite its own struct.
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit df2252054eb0f9f65389f3911588256bcc4f973b
Author: Ian Rogers <irogers@google.com>
Date: Wed Nov 3 23:41:49 2021 -0700

description
===========
By switching to an array of pointers to tests (later to be suites)
the definition of the tests can be moved to the file containing the
tests.

Committer notes:

It's "&vectors_page", not "&vectors_pages", noticed when cross building
to 32-bit ARM.

Also the DEFINE_SUITE(vectors_page) should be done where its function is
implemented, in tools/perf/arch/arm/tests/vectors-page.c, so that we can
make it static, as we don't have anymore its declaration in tests.h.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>