perf expr: Add source_count for aggregating events
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit 9aba0adae8c773ba0c0adc7c4d97768e044166cb
Author: Ian Rogers <irogers@google.com>
Date: Wed Nov 10 16:21:09 2021 -0800

description
===========
Events like uncore_imc/cas_count_read/ on Skylake open multiple events
and then aggregate in the metric leader. To determine the average value
per event the number of these events is needed. Add a source_count
function that returns this value by counting the number of events with
the given metric leader. For most events the value is 1 but for
uncore_imc/cas_count_read/ it can yield values like 6.

Add a generic test, but manually tested with a test metric that uses
the function.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>