perf arm-spe: Implement find_snapshot callback
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit 56c31cdff7c2a640f4afcfe0ac4e4ca6dc47c5fd
Author: German Gomez <german.gomez@arm.com>
Date: Tue Nov 9 16:30:08 2021 +0000

description
===========
The head pointer of the AUX buffer managed by the arm_spe_pmu.c driver
is not monotonically increasing, therefore the find_snapshot callback is
needed in order to find the trace data within the AUX buffer and avoid
wasting space in the perf.data file.

The pointer is assumed to have wrapped if the buffer contains non-zero
data at the end. If it has wrapped, the entire contents of the AUX
buffer are stored in the perf.data file. Otherwise only the data up to
the head pointer is stored.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>