perf tools: Prevent out-of-bounds access to registers
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit c897899752478d4c905c56f2b54b99ba82b34e13
Author: German Gomez <german.gomez@arm.com>
Date: Wed Dec 1 12:33:29 2021 +0000

description
===========
The size of the cache of register values is arch-dependant
(PERF_REGS_MAX). This has the potential of causing an out-of-bounds
access in the function "perf_reg_value" if the local architecture
contains less registers than the one the perf.data file was recorded on.

Since the maximum number of registers is bound by the bitmask "u64
cache_mask", and the size of the cache when running under x86 systems is
64 already, fix the size to 64 and add a range-check to the function
"perf_reg_value" to prevent out-of-bounds access.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>