perf intel-pt: Fix sync state when a PSB (synchronization) packet is found
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit ad106a26aef3a95ac7ca88d033b431661ba346ce
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Dec 10 18:22:58 2021 +0200

description
===========
When syncing, it may be that branch packet generation is not enabled at
that point, in which case there will not immediately be a control-flow
packet, so some packets before a control flow packet turns up, get
ignored.  However, the decoder is in sync as soon as a PSB is found, so
the state should be set accordingly.

Fixes: f4aa081949e7b6 ("perf tools: Add Intel PT decoder")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>