perf trace: Avoid early exit due to running SIGCHLD handler before it makes sense to
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit f06a82f9d31a87878a9295bac1defdadbc77bbc0
Author: Jiri Olsa <jolsa@redhat.com>
Date: Thu Jan 6 23:20:30 2022 +0100

description
===========
When running 'perf trace' with an BPF object like:

  # perf trace -e openat,tools/perf/examples/bpf/hello.c

the event parsing eventually calls llvm__get_kbuild_opts() that runs a
script and that ends up with SIGCHLD delivered to the 'perf trace'
handler, which assumes the workload process is done and quits 'perf
trace'.

Move the SIGCHLD handler setup directly to trace__run(), where the event
is parsed and the object is already compiled.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>