libtraceevent: Add __rel_loc relative location attribute support
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit cd7729043b31b32b34daaa6775e255b76de7f976
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date: Mon Nov 22 18:30:40 2021 +0900

description
===========
Add '__rel_loc' new dynamic data location attribute which encodes
the data location from the next to the field itself. This is similar
to the '__data_loc' but the location offset is not from the event
entry but from the next of the field.

This patch adds '__rel_loc' decoding support in the libtraceevent.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>