perf tools: Drop requirement for libstdc++.so for libopencsd check
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit ed17b1914978eddb2b01f2d34577f1c82518c650
Author: Uwe Kleine-König <uwe@kleine-koenig.org>
Date: Fri Dec 3 22:05:44 2021 +0100

description
===========
It's possible to link against libopencsd_c_api without having
libstdc++.so available, only libstdc++.so.6.0.28 (or whatever version is
in use) needs to be available. The same holds true for libopencsd.so.
When -lstdc++ (or -lopencsd) is explicitly passed to the linker however
the .so file must be available.

So wrap adding the dependencies into a check for static linking that
actually requires adding them all. The same construct is already used
for some other tests in the same file to reduce dependencies in the
dynamic linking case.

Fixes: 573cf5c9a152 ("perf build: Add missing -lstdc++ when linking with libopencsd")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>