perf bench: Use unbuffered output when pipe/tee'ing to a file
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit f0a29c9647ff8bbb424641f79bc1894e83dec218
Author: Sohaib Mohamed <sohaib.amhmd@gmail.com>
Date: Fri Nov 19 08:14:08 2021 +0200

description
===========
The output of 'perf bench' gets buffered when I pipe it to a file or to
tee, in such a way that I can see it only at the end.

E.g.

  $ perf bench internals synthesize -t
  < output comes out fine after each test run >

  $ perf bench internals synthesize -t | tee file.txt
  < output comes out only at the end of all tests >

This patch resolves this issue for 'bench' and 'test' subcommands.

See, also:

  $ perf bench mem all | tee file.txt
  $ perf bench sched all | tee file.txt
  $ perf bench internals all -t | tee file.txt
  $ perf bench internals all | tee file.txt

Committer testing:

It really gets staggered, i.e. outputs in bursts, when the buffer fills
up and has to be drained to make up space for more output.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>