perf vendor events: Rename arm64 arch std event files
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit 6732f10b11c63112ce34a064b247c00e090b9514
Author: Andrew Kilroy <andrew.kilroy@arm.com>
Date: Fri Dec 10 12:37:05 2021 +0000

description
===========
A previous commit adds pmu events into the files

  armv8-common-and-microarch.json
  armv8-recommended.json

that are actually specified in an armv9 reference supplement, not armv8.
As such, naming the files with the armv8 prefix seems artificial.

This patch renames the files to reflect that these two files are for
arch std events regardless of whether they are defined in armv8 or
armv9.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>