perf arm64: Rename perf_event_arm_regs for ARM64 registers
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit d3b58af9a8276c24a2aa80a059d87d99f5216d3b
Author: German Gomez <german.gomez@arm.com>
Date: Tue Dec 7 18:06:51 2021 +0000

description
===========
The registers for ARM and ARM64 are enumerated using two enums that have
the same name. In order to be able to import both headers, the name of
one can be replaced using the C preprocessor like so:

  #define perf_event_arm_regs perf_event_arm64_regs
  #include <asm/perf_regs.h>
  #undef perf_event_arm_regs

This patch updates all imports of ARM64's perf_regs.h in order to
prevent the naming collision.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>