perf ftrace: Move out common code from __cmd_ftrace
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit a9b8ae8ae347941fefd6596f62586b13ae032e4b
Author: Namhyung Kim <namhyung@kernel.org>
Date: Wed Dec 15 10:51:51 2021 -0800

description
===========
The signal setup code and evlist__prepare_workload() can be used for
other subcommands.  Let's move them out of the __cmd_ftrace().  Then
it doesn't need to pass argc and argv.

On the other hand, select_tracer() is specific to the 'trace'
subcommand so it'd better moving it into the __cmd_ftrace().

Signed-off-by: Michael Petlan <mpetlan@redhat.com>