perf evsel: Improve error message for uncore events
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit dcffc5ebb80dd5887b91091b8ecd082c9ed75361
Author: Ian Rogers <irogers@google.com>
Date: Thu Dec 23 10:39:48 2021 -0800

description
===========
When a group has multiple events and the leader fails it can yield
errors like:

  $ perf stat -e '{uncore_imc/cas_count_read/},instructions' /bin/true
  Error:
  The sys_perf_event_open() syscall returned with 22 (Invalid argument) for event (uncore_imc/cas_count_read/).
  /bin/dmesg | grep -i perf may provide additional information.

However, when not the group leader <not supported> is given:

  $ perf stat -e '{instructions,uncore_imc/cas_count_read/}' /bin/true
  ...
           1,619,057      instructions
     <not supported> MiB  uncore_imc/cas_count_read/

This is necessary because get_group_fd will fail if the leader fails and
is the direct result of the check on line 750 of builtin-stat.c in
stat_handle_error that returns COUNTER_SKIP for the latter case.

This patch improves the error message to:

  $ perf stat -e '{uncore_imc/cas_count_read/},instructions' /bin/true
  Error:
  Invalid event (uncore_imc/cas_count_read/) in per-thread mode, enable system wide with '-a'.

v2. Changed the test to use !target__has_cpu as suggested by Namhyung Kim.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>