perf cpumap: Add CPU to aggr_cpu_id
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit 34794913e2dc08a464499f795073a021feeb3b47
Author: Ian Rogers <irogers@google.com>
Date: Tue Jan 4 22:13:27 2022 -0800

description
===========
With no aggregration, such as 'perf stat -A', the aggr_cpu_id lacks a
way to describe per CPU aggregation and the core is set to the CPU in
places like print_counter_aggrdata in stat-display.c.

Setting the core to the CPU is undesirable as the CPU will exceed valid
core values and lead to confusion. Add a CPU variable to address this.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>