perf evsel: Derive CPUs and threads in alloc_counts
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit 2ca0a3718da24953689b1771589ac63b60f17358
Author: Ian Rogers <irogers@google.com>
Date: Tue Jan 4 22:13:29 2022 -0800

description
===========
Passing the number of CPUs and threads allows for an evsel's counts to
be mismatched to its cpu map. To avoid this always derive the counts
size from the cpu map. Change openat-syscall-all-cpus to set the cpus
to allow for this to work.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>