perf cpumap: Add is_dummy() method
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit 37be585807cb9a810f8395c39c4ee7bdbdc7b0dc
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Sat Jan 15 17:11:10 2022 -0300

description
===========
Needed to check if a cpu_map is dummy, i.e. not a cpu map at all, for
pid monitoring scenarios.

This probably needs to move to libperf, but since perf itself is the
first and so far only user, leave it at tools/perf/util/.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>