perf stat: No need to setup affinities when starting a workload
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit 49de179577e7b05b57f625bf05cdc60a72de38d0
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Jan 17 13:09:29 2022 -0300

description
===========
I.e. the simple:

  $ perf stat sleep 1

Uses a dummy CPU map and thus there is no need to setup/cleanup
affinities to avoid IPIs, etc.

With this we're down to a sched_getaffinity() call, in the libnuma
initialization, that probably can be removed in a followup patch.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>