perf test: Skip failing sigtrap test for arm+aarch64
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit f268088f65af69aa6ae1fb65e696cbb6478bcc9a
Author: John Garry <john.garry@huawei.com>
Date: Fri Feb 18 17:33:33 2022 +0800

description
===========
Skip the Sigtrap test for arm + arm64, same as was done for s390 in
commit a840974e96fd ("perf test: Test 73 Sig_trap fails on s390"). For
this, reuse BP_SIGNAL_IS_SUPPORTED - meaning that the arch can use BP to
generate signals - instead of BP_ACCOUNT_IS_SUPPORTED, which is
appropriate.

As described by Will at [0], in the test we get stuck in a loop of
handling the HW breakpoint exception and never making progress. GDB
handles this by stepping over the faulting instruction, but with perf
the kernel is expected to handle the step (which it doesn't for arm).

Dmitry made an attempt to get this work, also mentioned in the same
thread as [0], which was appreciated. But the best thing to do is skip
the test for now.

[0] https://lore.kernel.org/linux-perf-users/20220118124343.GC98966@leoy-ThinkPad-X240s/T/#m13b06c39d2a5100d340f009435df6f4d8ee57b5a

Fixes: 5504f67944484495 ("perf test sigtrap: Add basic stress test for sigtrap handling")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>