ACPI / x86: Drop PWM2 device on Lenovo Yoga Book from always present table
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2067294

commit d431dfb764b145369be820fcdfd50f2159b9bbc2
Author: Hans de Goede <hdegoede@redhat.com>
Date: Mon, 22 Nov 2021 18:05:30 +0100

It turns out that there is a WMI object which controls the PWM2 device
used for the keyboard backlight and that WMI object also provides some
other useful functionality.

The upcoming lenovo-yogabook-wmi driver will offer both backlight
control and the other functionality, so there no longer is a need
to have the lpss-pwm driver binding to PWM2 for backlight control;
and this is now actually undesirable because this will cause both
the WMI code and the lpss-pwm driver to poke at the same PWM
controller.

Drop the always-present quirk for the PWM2 ACPI-device, so that the
 lpss-pwm controller will no longer bind to it.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>