ACPI: EC: Mark the ec_sys write_support param as module_param_hw()
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2067294

commit b66f86849414807745b5c2129e2de5f27a788c9f
Author: Hans de Goede <hdegoede@redhat.com>
Date: Thu, 25 Nov 2021 11:36:16 +0100

Using write_support=1 with the ec_sys module changes the mode of the
"io" debugfs file to 0600. This will cause any attempts to access it under
a kernel in lockdown mode to return -EPERM, which makes the entire ec_sys
module unusable.

Use the special module_param_hw() macro for module parameters which
may not be used while in lockdown mode, to avoid this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>