ACPICA: Fix AEST Processor generic resource substructure data field byte length
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2067294

commit 00395b74d57ff81795ec392627db7e1764c94941
Author: Shuuichirou Ishii <ishii.shuuichir@fujitsu.com>
Date: Wed, 22 Dec 2021 17:28:34 +0100

ACPICA commit 13b9327761955f6e1e5dbf748b3112940c0dc539

The byte length of the Data field in the AEST Processor generic resource
substructure defined in ACPI for the Armv8 RAS Extensions 1.1 is 4Byte.
However, it is defined as a pointer type, and on a 64-bit machine,
it is interpreted as 8 bytes. Therefore, it is changed from a pointer
type unsigned integer 1 byte to an unsigned integer 4 bytes.

Link: https://github.com/acpica/acpica/commit/13b93277
Signed-off-by: Shuuichirou Ishii <ishii.shuuichir@fujitsu.com>
Signed-off-by: Bob Moore <robert.moore@intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>