perf unwind: Don't show unwind error messages when augmenting frame pointer stack
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit fa7095c5c3240bb2ecbc77f8b69be9b1d9e2cf60
Author: James Clark <james.clark@arm.com>
Date:   Wed Apr 6 15:56:51 2022 +0100

description
===========
Commit Fixes: b9f6fbb3b2c29736 ("perf arm64: Inject missing frames when
using 'perf record --call-graph=fp'") intended to add a 'best effort'
DWARF unwind that improved the frame pointer stack in most scenarios.

It's expected that the unwind will fail sometimes, but this shouldn't be
reported as an error. It only works when the return address can be
determined from the contents of the link register alone.

Fix the error shown when the unwinder requires extra registers by adding
a new flag that suppresses error messages. This flag is not set in the
normal --call-graph=dwarf unwind mode so that behavior is not changed.

Fixes: b9f6fbb3b2c29736 ("perf arm64: Inject missing frames when using 'perf record --call-graph=fp'")

Link: https://lore.kernel.org/r/20220406145651.1392529-1-james.clark@arm.com

Signed-off-by: Michael Petlan <mpetlan@redhat.com>