perf/x86: Add new event for AUX output counter index
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit 8b8ff8cc3b8155c18162e8b1f70e1230db176862
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Sep 7 19:39:01 2021 +0300

description
===========
PEBS-via-PT records contain a mask of applicable counters. To identify
which event belongs to which counter, a side-band event is needed. Until
now, there has been no side-band event, and consequently users were limited
to using a single event.

Add such a side-band event. Note the event is optimised to output only
when the counter index changes for an event. That works only so long as
all PEBS-via-PT events are scheduled together, which they are for a
recording session because they are in a single group.

Also no attribute bit is used to select the new event, so a new
kernel is not compatible with older perf tools.  The assumption
being that PEBS-via-PT is sufficiently esoteric that users will not
be troubled by this.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>