perf tests: Fix coresight `perf test` failure.
Bugzilla: https://bugzilla.redhat.com/2069073

upstream
========
commit 45fa7c38696bae632310c2876ba81fdfa25cc9c2
Author: Jeremy Linton <jeremy.linton@arm.com>
Date:   Thu Apr 28 10:19:47 2022 -0500

description
===========
Currently the `perf test` always fails the coresight test like:

  89: Check Arm CoreSight trace data recording and synthesized samples: FAILED!

That is because the test_arm_coresight.sh is attempting to SIGINT the
parent but is using $$ rather than $PPID and it sigint's itself when
run under the perf test framework.

Since this is done in a trap clause it ends up returning a non zero
return.

Since $PPID is a bash ism and not all distros are linking /bin/sh to
bash, the alternative parent pid lookups are uglier than just dropping
the kill, and its not strictly needed, lets pick the simple solution and
drop the sigint.

Fixes: 133fe2e617e48ca0 ("perf tests: Improve temp file cleanup in test_arm_coresight.sh")

Link: https://lore.kernel.org/r/20220428151947.290146-1-jeremy.linton@arm.com

Signed-off-by: Michael Petlan <mpetlan@redhat.com>