ACPI / x86: Add support for LPS0 callback handler
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2067297

commit 20e1d6402a71dba7ad2b81f332a3c14c7d3b939b
Author: Mario Limonciello <mario.limonciello@amd.com>
Date: Thu, 17 Mar 2022 09:14:42 -0500

Currenty the latest thing run during a suspend to idle attempt is
the LPS0 `prepare_late` callback and the earliest thing is the
`resume_early` callback.

There is a desire for the `amd-pmc` driver to suspend later in the
suspend process (ideally the very last thing), so create a callback
that it or any other driver can hook into to do this.

Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Acked-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Link: https://lore.kernel.org/r/20220317141445.6498-1-mario.limonciello@amd.com
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>