ACPI: IPMI: replace usage of found with dedicated list iterator variable
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2067297

commit 26de0ab9841a1610a477d9b07c024b9bb6d39f98
Author: Jakob Koschel <jakobkoschel@gmail.com>
Date: Thu, 24 Mar 2022 08:04:41 +0100

To move the list iterator variable into the list_for_each_entry_*()
macro in the future it should be avoided to use the list iterator
variable after the loop body.

To *never* use the list iterator variable after the loop it was
concluded to use a separate iterator variable instead of a
found boolean [1].

This removes the need to use a found variable and simply checking if
the variable was set, can determine if the break/goto was hit.

Link: https://lore.kernel.org/all/CAHk-=wgRr_D8CB-D9Kg-c=EHreAsk5SqXPwr9Y7k9sA6cWXJ6w@mail.gmail.com/
Signed-off-by: Jakob Koschel <jakobkoschel@gmail.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>