ASoC: tegra20-spdif: stop setting slave_id
The DMA resource is never set up anywhere, and passing this as slave_id
has not been the proper procedure in a long time.

As a preparation for removing all slave_id references from the ALSA code,
remove this one.

According to Dmitry Osipenko, this driver has never been used and
the mechanism for configuring DMA would not work as it is implemented,
so this part will get rewritten when the driver gets put into use
again in the future.

    Signed-off-by: Arnd Bergmann <arnd@arndb.de>
    Reviewed-by: Dmitry Osipenko <digetx@gmail.com>
    Acked-by: Mark Brown <broonie@kernel.org>
    Link: https://lore.kernel.org/r/20211122222203.4103644-2-arnd@kernel.org
    Signed-off-by: Vinod Koul <vkoul@kernel.org>

Author: Arnd Bergmann <arnd@arndb.de>
Date: Mon Nov 22 23:21:53 2021 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 0725ac9ac4492568514a94971f46dd0546684ff8)
Bugzilla: https://bugzilla.redhat.com/2065572