ASoC: rt5682: Register wclk with its parent_hws instead of parent_data
The mclk might not be registered as a fixed clk name "mclk" on some
platforms.
In those platforms, if the mclk needed to be controlled by codec driver
and acquired by a fixed name, it would be a problem.

This patch to fix the issue that wclk becomes an orphan due to the fixed
mclk's name.

    Signed-off-by: Derek Fang <derek.fang@realtek.com>
    Link: https://lore.kernel.org/r/20211227055446.27563-1-derek.fang@realtek.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Derek Fang <derek.fang@realtek.com>
Date: Mon Dec 27 13:54:46 2021 +0800

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit cc5c9788106fb1b9e03c8c57d8d7166073a54416)
Bugzilla: https://bugzilla.redhat.com/2065572