ASoC: Intel: bytcr_rt5640: Add support for external GPIO jack-detect
Some boards have the codec IRQ hooked-up as normally, so the driver can
still do things like headset vs headphones and button-press detection,
but instead of using one of the JD pins of the codec, an external GPIO
is used to report the jack-presence switch status of the jack.

Add support for boards which have this setup and which specify which
external GPIO to use in the special Android AMCR0F28 ACPI device.

And add a quirk for the Asus TF103C tablet which uses this setup.

    Signed-off-by: Hans de Goede <hdegoede@redhat.com>
    Link: https://lore.kernel.org/r/20220106110128.66049-7-hdegoede@redhat.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Hans de Goede <hdegoede@redhat.com>
Date: Thu Jan 6 12:01:28 2022 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 44125fd5315154c6b8326b5c27646af3b33ba25c)
Bugzilla: https://bugzilla.redhat.com/2065572