ASoC: simple-card-utils: Add new system-clock-fixed flag
Add a new system-clock-fixed flag, which can be used to specify that the
driver cannot or should not allow the clock frequency of the mapped clock
to be modified. This behavior is also implied if the system-clock-frequency
parameter is set explicitly - the flag is meant for cases where a clock is
mapped to the DAI but which is, or should be treated as, fixed.

When mclk-fs is also specified, this causes a PCM constraint to be added
which enforces that only the corresponding valid sample rate can be used.

    Signed-off-by: Robert Hancock <robert.hancock@calian.com>
    Link: https://lore.kernel.org/r/20220120195832.1742271-7-robert.hancock@calian.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Robert Hancock <robert.hancock@calian.com>
Date: Thu Jan 20 13:58:32 2022 -0600

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 5ca2ab4598179a2690a38420f3fde9f2ad79d55c)
Bugzilla: https://bugzilla.redhat.com/2065572