ASoC: sun4i-i2s: Update registers for more channels
H6 expands the number of channels in each direction to 16, so the slot
number fields need to be expanded from 3 to 4 bits each.

R329/D1 expand that further by allowing each of the 16 slots to map to
any of 4 data pins. For TX, the configuration of each pin is
independent, so there is a copy of the mapping registers for each pin.
For RX, each of the 16 slots can map to only one pin, so the registers
were changed to add the pin selection inline with the channel mapping.

    Signed-off-by: Samuel Holland <samuel@sholland.org>
    Acked-by: Maxime Ripard <maxime@cerno.tech>
    Link: https://lore.kernel.org/r/20220203020116.12279-3-samuel@sholland.org
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Samuel Holland <samuel@sholland.org>
Date: Wed Feb 2 20:01:14 2022 -0600

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit c8bbc1de9088fedb5d71db7d185c37db18feb2e1)
Bugzilla: https://bugzilla.redhat.com/2065572