ASoC: codecs: Add Awinic AW8738 audio amplifier driver
The Awinic AW8738 is a simple audio amplifier using a single GPIO.
The main difference to simple-amplifier is that there is a "one-wire
pulse control" that allows configuring the amplifier to one of a few
pre-defined modes. This can be used to configure the speaker-guard
function (primarily the power limit for the amplifier).

Add a simple driver that allows setting it up in the device tree
with a specified mode number.

    Signed-off-by: Jonathan Albrieux <jonathan.albrieux@gmail.com>
    Co-developed-by: Stephan Gerhold <stephan@gerhold.net>
    Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
    Link: https://lore.kernel.org/r/20220304102452.26856-3-stephan@gerhold.net
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Jonathan Albrieux <jonathan.albrieux@gmail.com>
Date: Fri Mar 4 11:24:52 2022 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 6b4528b5532f84f385b4e756637698cf4ae211f3)
Bugzilla: https://bugzilla.redhat.com/2065572