[redhat] New configs in drivers/net/can
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CAN_CTUCANFD_PCI [=n]
 Type  : tristate
 Defined at drivers/net/can/ctucanfd/Kconfig:14
   Prompt: CTU CAN-FD IP core PCI/PCIe driver
   Depends on: NET [=y] && CAN [=m] && CAN_DEV [=m] && PCI [=y]
   Location:
     Main menu
       -> Networking support (NET [=y])
         -> CAN bus subsystem support (CAN [=m])
           -> CAN Device Drivers
             -> Platform CAN drivers with Netlink support (CAN_DEV [=m])
 Selects: CAN_CTUCANFD [=n]

---

 Symbol: CAN_CTUCANFD_PLATFORM [=n]
 Type  : tristate
 Defined at drivers/net/can/ctucanfd/Kconfig:24
   Prompt: CTU CAN-FD IP core platform (FPGA, SoC) driver
   Depends on: NET [=y] && CAN [=m] && CAN_DEV [=m] && HAS_IOMEM [=y] && (OF [=y] || COMPILE_TEST [=n])
   Location:
     Main menu
       -> Networking support (NET [=y])
         -> CAN bus subsystem support (CAN [=m])
           -> CAN Device Drivers
             -> Platform CAN drivers with Netlink support (CAN_DEV [=m])
 Selects: CAN_CTUCANFD [=n]

---

Cc: Peter Robinson <pbrobinson@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>